+++ crates/affinidi-did-resolver/affinidi-did-resolver-methods/did-webvh/src/log_entry/mod.rs
@@ -1,8 +1,10 @@
 /*!
 *   Webvh utilizes Log Entries for each version change of the DID Document.
 */
-use crate::{DIDWebVHError, parameters::Parameters};
-use affinidi_data_integrity::{DataIntegrityProof, SignedDocument, SigningDocument};
+use crate::{DIDWebVHError, parameters::Parameters, witness::Witnesses};
+use affinidi_data_integrity::{
+    DataIntegrityProof, SignedDocument, SigningDocument, verification_proof::verify_data,
+};
 use multibase::Base;
 use multihash::Multihash;
 use serde::{Deserialize, Serialize};
@@ -12,8 +14,23 @@ use sha2::{Digest, Sha256};
 use std::{collections::HashMap, fs::OpenOptions, io::Write};
 use tracing::debug;
 
-pub mod create;
 pub mod read;
+
+/// Resolved Document MetaData
+/// Returned as reolved Document MetaData on a successful resolve
+#[derive(Clone, Debug, Default, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct MetaData {
+    pub version_id: String,
+    pub version_time: String,
+    pub created: String,
+    pub updated: String,
+    pub scid: String,
+    pub portable: bool,
+    pub deactivated: bool,
+    pub witness: Option<Witnesses>,
+    pub watchers: Option<Vec<String>>,
+}
 
 /// Each version of the DID gets a new log entry
 /// [Log Entries](https://identity.foundation/didwebvh/v1.0/#the-did-log-file)
@@ -76,7 +93,7 @@ impl LogEntry {
 
     /// Generates a SCID from a preliminary LogEntry
     /// This only needs to be called once when the DID is first created.
-    fn generate_scid(&self) -> Result<String, DIDWebVHError> {
+    pub(crate) fn generate_scid(&self) -> Result<String, DIDWebVHError> {
         self.generate_log_entry_hash().map_err(|e| {
             DIDWebVHError::SCIDError(format!(
                 "Couldn't generate SCID from preliminary LogEntry. Reason: {}",
@@ -105,6 +122,53 @@ impl LogEntry {
         })?;
 
         Ok(multibase::encode(Base::Base58Btc, hash_encoded.to_bytes()))
+    }
+
+    pub fn validate_witness_proof(
+        &self,
+        witness_proof: &DataIntegrityProof,
+    ) -> Result<bool, DIDWebVHError> {
+        // Create a SigningDocument from the LogEntry
+        let mut signing_doc: SignedDocument = self.try_into()?;
+        signing_doc.proof = Some(witness_proof.clone());
+        signing_doc.extra.insert(
+            "proof".to_string(),
+            serde_json::to_value(&self.proof).map_err(|e| {
+                DIDWebVHError::ParametersError(format!(
+                    "Couldn't serialize LogEntry Proof to JSON Value: {}",
+                    e
+                ))
+            })?,
+        );
+
+        // Verify the Data Integrity Proof against the Signing Document
+        verify_data(&signing_doc).map_err(|e| {
+            DIDWebVHError::LogEntryError(format!("Data Integrity Proof verification failed: {}", e))
+        })?;
+
+        Ok(true)
+    }
+
+    /// Splits the version number and the version hash for a DID versionId
+    pub fn get_version_id_fields(&self) -> Result<(u32, String), DIDWebVHError> {
+        LogEntry::parse_version_id_fields(&self.version_id)
+    }
+
+    /// Splits the version number and the version hash for a DID versionId
+    pub fn parse_version_id_fields(version_id: &str) -> Result<(u32, String), DIDWebVHError> {
+        let Some((id, hash)) = version_id.split_once('-') else {
+            return Err(DIDWebVHError::ValidationError(format!(
+                "versionID ({}) doesn't match format <int>-<hash>",
+                version_id
+            )));
+        };
+        let id = id.parse::<u32>().map_err(|e| {
+            DIDWebVHError::ValidationError(format!(
+                "Failed to parse version ID ({}) as u32: {}",
+                id, e
+            ))
+        })?;
+        Ok((id, hash.to_string()))
     }
 }