+++ crates/affinidi-did-resolver/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ## Changelog history
 
+### 26th June 2025
+
+## DID webvh method (0.1.5)
+
+* **CHANGE:** `DIDWebVHState` is now the entry point to working with webvh
+LogEntries and WitnessProofs
+* **FEATURE:** Witness Proofs are now optimized based on webvh 1.0 spec
+  * Prior witness entries will be removed, only keeping the latest witness proof
+  for each witness
+* **FEATURE:** Witness proof validation enabled
+  * Correct handling of optimised Witness Proofs is managed by `WitnessProofCollection`
+* **FEATURE:** Future witness proofs will be ignored based on highest versionId
+from LogEntry
+
 ### 19th June 2025
 
 ## DID webvh method (0.1.4)